Integrated template matching
Fully (more-or-less) integrated template matching into PeakFinder class. Need
to do a lot of cleaning up of match_temp, as currently it runs ~3 seconds for
a very short set of data. But, should be functional and is capable of
detecting missing peaks, as expected.

Also, made a version update even though the versions are meaningless! We're
calling this v0.1, because it technically functions but is a sad, sad shell of
a module. More to come!