Modality updates
Updated some defaults in modalities for detecting peaks and manual peak editing.

Also added a peakdet.save() and peakdet.load() method for pickling classes. This
desperately needs to be updated, but will require some additional methods for
PeakFinder to store history of filtering/interpolation commands and output a
JSON file instead of a pickle (which, on test, ends up taking ~15x more disk
space than the raw data). But, for now, it's a placeholder to remind me to write
something else.