[REF] Fix allow_pickle issue in numpy.load
This is a bandaid and `np.load` and `np.savez_compressed` should be
removed from `peakdet.io` and replaced with something more sustainable
(like saving the data as a JSON file or something). OR just removed
altogether, if necessary. Pickling SHOULD NOT be allowed, but I just
want to get tests passing so that I can transfer ownership of this thing
to physiopy